Module restructuring, fix Math interpreter should be in a seperate module #17
